% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_status}
\alias{check_status}
\title{Check Request Status From Twitch}
\usage{
check_status(response)
}
\arguments{
\item{response}{A response objct. Generally generated from \code{httr::GET()} or \code{httr::POST()}.}
}
\value{
An error message if status code is not 200. Invisible \code{NULL} if status code is 200.
}
\description{
Check Request Status From Twitch
}
