% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_bits_cheermotes}
\alias{clean_bits_cheermotes}
\title{Clean the Response From the Clean Bits Cheermotes Endpoint}
\usage{
clean_bits_cheermotes(response_content)
}
\arguments{
\item{response_content}{Content from a HTTP Request. This is expecting to take the output from \code{httr::content()}.}
}
\value{
Clean and tidy tibble.
}
\description{
Clean the Response From the Clean Bits Cheermotes Endpoint
}
