% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_get_schedule}
\alias{clean_get_schedule}
\title{Clean the Response From User Follows}
\usage{
clean_get_schedule(response_content)
}
\arguments{
\item{response_content}{Content from a HTTP Request. This is expecting to take the output from \code{httr::content()}.}
}
\value{
A named list containing: data, a Clean and tidy tibble. pagination, cursor for pagination.
}
\description{
Clean the Response From User Follows
}
