% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_users}
\alias{clean_users}
\title{Clean the Response from Users}
\usage{
clean_users(response_content)
}
\arguments{
\item{response_content}{Content from a HTTP Request. This is expecting to take the output from \code{httr::content()}.}
}
\value{
Clean and tidy tibble.
}
\description{
Clean the Response from Users
}
