% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{date_formatter}
\alias{date_formatter}
\title{Wrapper for Formatting Date Times}
\usage{
date_formatter(.data)
}
\arguments{
\item{.data}{A tibble. Data to be formatted.}
}
\description{
Wrapper for Formatting Date Times
}
