% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{flatten_keep_names}
\alias{flatten_keep_names}
\title{Keep names while flattening a named list}
\usage{
flatten_keep_names(x, use.names = TRUE, classes = "ANY")
}
\arguments{
\item{x}{\code{list}}

\item{use.names}{\code{logical}. Should the names of \code{x} be kept?}

\item{classes}{A character vector of class names, or "ANY" to match any class.}
}
\description{
Source taken from \code{rlist::list.flatten}. Didn't want to load dependency. Added a step to clean the names to snake case.
}
