% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{get_all_follows}
\alias{get_all_follows}
\title{Get All Follows}
\usage{
get_all_follows(from_id = NULL, to_id = NULL)
}
\arguments{
\item{from_id}{A character. User ID. The request returns information about users who are being followed by the from_id user.}

\item{to_id}{A character. User ID. The request returns information about users who are following the to_id user.}
}
\description{
Get All Follows
}
\examples{
\dontrun{
twitch_auth()

user <- get_users(login = "TheEatGameLove")

all_followers <- get_all_follows(to_id = user$id)
}
}
\references{
https://dev.twitch.tv/docs/api/reference#get-clips
}
\seealso{
Other Users: 
\code{\link{get_chatters}()},
\code{\link{get_follows}()},
\code{\link{get_users}()}
}
\concept{Users}
