% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/videos.R
\name{get_all_videos}
\alias{get_all_videos}
\title{Get All Follows}
\usage{
get_all_videos(
  id = NULL,
  user_id = NULL,
  game_id = NULL,
  language = NULL,
  period = NULL,
  sort = NULL,
  type = NULL,
  clean_json = TRUE
)
}
\arguments{
\item{id}{A string. Required (Read Documentation for more information). ID of the video being queried. Limit: 100. If this is specified, you cannot use any of the optional query parameters below.}

\item{user_id}{A string. Required (Read Documentation for more information). ID of the user who owns the video. Limit 1.}

\item{game_id}{A string. Required (Read Documentation for more information). ID of the game the video is of. Limit 1.}

\item{language}{A string. Optional.     Language of the video being queried. Limit: 1.}

\item{period}{A string. Optional. Period during which the video was created. Valid values: "all", "day", "week", "month". Default: "all".}

\item{sort}{A string. Optional. Sort order of the videos. Valid values: "time", "trending", "views". Default: "time".}

\item{type}{A string. Optional. ype of video. Valid values: "all", "upload", "archive", "highlight". Default: "all".}

\item{clean_json}{A logical. If \code{TRUE}, clean and tidy the data. If \code{FALSE}, return the result of \code{httr::content}.}
}
\description{
Get All Follows
}
\examples{
\dontrun{
library(twitchr)

twitch_auth()

all_videos <- get_all_videos(user_id = 613890167)
}
}
\references{
https://dev.twitch.tv/docs/api/reference#get-videos
}
\seealso{
Other Users: 
\code{\link{get_all_follows}()},
\code{\link{get_chatters}()},
\code{\link{get_follows}()},
\code{\link{get_schedule}()},
\code{\link{get_users}()}

Other Videos: 
\code{\link{get_videos}()}
}
\concept{Users}
\concept{Videos}
