% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{get_chatters}
\alias{get_chatters}
\title{Get A Broadcasters Current Chatters}
\usage{
get_chatters(broadcaster)
}
\arguments{
\item{broadcaster}{A string. The name of the broadcaster.}
}
\value{
A tibble data frame of chatter data.
}
\description{
\strong{This uses undocumented API endpoints. This may change or break at anytime. Use at your own risk!}.
}
\examples{
\dontrun{
library(twitchr)

twitch_auth()

chatters <- get_chatters("KowAndToilet")
}
}
\references{
https://twitch.uservoice.com/forums/310213-developers/suggestions/39145294-chatters-viewers-helix-api-endpoint
}
\seealso{
Other Users: 
\code{\link{get_all_follows}()},
\code{\link{get_all_videos}()},
\code{\link{get_follows}()},
\code{\link{get_schedule}()},
\code{\link{get_users}()}
}
\concept{Users}
