% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cheermotes.R
\name{get_clips}
\alias{get_clips}
\title{Get Clips}
\usage{
get_clips(broadcaster_id = NULL, game_id = NULL, id = NULL, clean_json = TRUE)
}
\arguments{
\item{broadcaster_id}{A character. ID of the broadcaster for whom clips are returned. The number of clips returned is determined by the first query-string parameter (default: 20). Results are ordered by view count.}

\item{game_id}{A character. ID of the game for which clips are returned. The number of clips returned is determined by the first query-string parameter (default: 20). Results are ordered by view count.}

\item{id}{A character. ID of the clip being queried. Limit: 100.}

\item{clean_json}{A logical. If \code{TRUE}, clean and tidy the data. If \code{FALSE}, return the result of \code{httr::content}.}
}
\description{
Gets clip information by clip ID (one or more), broadcaster ID (one only), or game ID (one only). Note: The clips service returns a maximum of 1000 clips.
}
\references{
https://dev.twitch.tv/docs/api/reference#get-clips
}
