% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{get_follows}
\alias{get_follows}
\title{Get Follows}
\usage{
get_follows(
  from_id = NULL,
  to_id = NULL,
  after = NULL,
  first = NULL,
  clean_json = TRUE
)
}
\arguments{
\item{from_id}{A character. User ID. The request returns information about users who are being followed by the from_id user.}

\item{to_id}{A character. User ID. The request returns information about users who are following the to_id user.}

\item{after}{A character. Cursor for forward pagination: tells the server where to start fetching the next set of results, in a multi-page response. The cursor value specified here is from the pagination response field of a prior query.}

\item{first}{A numeric. Maximum number of objects to return. Maximum: 100. Default: 20.}

\item{clean_json}{A logical. If \code{TRUE}, clean and tidy the data. If \code{FALSE}, return the result of \code{httr::content}.}
}
\value{
A tibble data frame of follower data.
}
\description{
Get Follows
}
\examples{
\dontrun{
library(twitchr)

twitch_auth()

user <- get_users(login = "TheEatGameLove")

followers <- get_follows(to_id = user$id)
}
}
\references{
https://dev.twitch.tv/docs/api/reference#get-users-follows
}
\seealso{
Other Users: 
\code{\link{get_all_follows}()},
\code{\link{get_chatters}()},
\code{\link{get_users}()}
}
\concept{Users}
