% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games.R
\name{get_games}
\alias{get_games}
\title{Get games}
\usage{
get_games(id = NULL, name = NULL, clean_json = TRUE)
}
\arguments{
\item{id}{A numeric. Game ID. At most 100 id values can be specified.}

\item{name}{A numeric. Game name. The name must be an exact match. For example, “Pokemon” will not return a list of Pokemon games; instead, query any specific Pokemon games in which you are interested. At most 100 name values can be specified.}

\item{clean_json}{A logical. If \code{TRUE}, clean and tidy the data. If \code{FALSE}, return the result of \code{httr::content}.}
}
\value{
A tibble data frame of game data.
}
\description{
Gets game information by game ID or name. For a query to be valid, \code{name} and/or \code{id} must be specified.
}
\examples{
\dontrun{
library(twitchr)

twitch_auth()

games <- get_games(name = c("Battletoads", "Stardew Valley"))
}
}
\references{
\url{https://dev.twitch.tv/docs/api/reference#get-games}
}
\seealso{
Other Games: 
\code{\link{get_top_games}()}
}
\concept{Games}
