% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{get_schedule}
\alias{get_schedule}
\title{Get Schedule}
\usage{
get_schedule(
  broadcaster_id = NULL,
  id = NULL,
  start_time = NULL,
  utc_offset = NULL,
  first = NULL,
  after = NULL,
  clean_json = TRUE
)
}
\arguments{
\item{broadcaster_id}{A string. User ID of the broadcaster who owns the channel streaming schedule. Provided broadcaster_id must match the user_id in the user OAuth token. Maximum: 1}

\item{id}{A string. The ID of the stream segment to return. Maximum: 100.}

\item{start_time}{A string. A timestamp in RFC3339 format to start returning stream segments from. If not specified, the current date and time is used.}

\item{utc_offset}{A string. A timezone offset for the requester specified in minutes. This is recommended to ensure stream segments are returned for the correct week. For example, a timezone that is +4 hours from GMT would be “240.” If not specified, “0” is used for GMT.}

\item{first}{A numeric. Maximum number of stream segments to return. Maximum: 25. Default: 20.}

\item{after}{A string. Cursor for forward pagination: tells the server where to start fetching the next set of results in a multi-page response. The cursor value specified here is from the pagination response field of a prior query.}

\item{clean_json}{A logical. If \code{TRUE}, clean and tidy the data. If \code{FALSE}, return the result of \code{httr::content}.}
}
\value{
A tibble data frame of a user's schedule data.
}
\description{
Gets all scheduled broadcasts or specific scheduled broadcasts from a channel’s \href{https://help.twitch.tv/s/article/channel-page-setup?language=en_US#Schedule}{stream schedule}. Scheduled broadcasts are defined as “stream segments” in the API.
}
\details{
If a Twitch streamer uses recurring scheduling the pagination will be infinite.
\subsection{Vacations}{

If a Twitch streamer uses the vacation feature on twitch the return will have "Vacation" as a \code{category_name} with a \code{category_id} of 0 for filtering + displaying vacation times.
}
}
\examples{
\dontrun{
library(twitchr)

twitch_auth()

users <- get_users(login = "TheEatGameLove")

schedule <- get_schedule(users$id)
}
}
\references{
https://dev.twitch.tv/docs/api/reference#get-channel-stream-schedule
}
\seealso{
Other Users: 
\code{\link{get_all_follows}()},
\code{\link{get_chatters}()},
\code{\link{get_follows}()},
\code{\link{get_users}()}
}
\concept{Users}
