% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{get_users}
\alias{get_users}
\title{Get users}
\usage{
get_users(id = NULL, login = NULL, clean_json = TRUE)
}
\arguments{
\item{id}{A character. Optional. User ID. Multiple user IDs can be specified. Limit: 100.}

\item{login}{A character. Optional. User login name. Multiple login names can be specified. Limit: 100.}

\item{clean_json}{A logical. If \code{TRUE}, clean and tidy the data. If \code{FALSE}, return the result of \code{httr::content}.}
}
\value{
A tibble data frame of twitch user data.
}
\description{
Gets information about one or more specified Twitch users. Users are identified by optional user IDs and/or login name. If neither a user ID nor a login name is specified, the user is looked up by Bearer token. \emph{Note: This package does not support the bearer token at this time.}
}
\examples{
\dontrun{
library(twitchr)

twitch_auth()

users <- get_users(login = "KowAndToilet")
}
}
\references{
\url{https://dev.twitch.tv/docs/api/reference#get-users}
}
\seealso{
Other Users: 
\code{\link{get_all_follows}()},
\code{\link{get_follows}()},
\code{\link{get_schedule}()}
}
\concept{Users}
