% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidden_endpoints.R
\name{get_video_comments}
\alias{get_video_comments}
\title{Get video comments}
\usage{
get_video_comments(video_id)
}
\arguments{
\item{video_id}{A numeric. Video ID for the comments to extract.}
}
\value{
A list of comments related to the requested \code{video_id}. \code{NULL} is returned if the \code{video_id} does not exist.
}
\description{
\strong{This uses undocumented API endpoints. This may change or break at anytime. Use at your own risk!}. Given a \code{video_id}, return the comments/chat related to that video.
}
\examples{
\dontrun{
library(twitchr)

twitch_auth()

video_comments <- get_video_comments(822494395)
}
}
\seealso{
Other Hidden Endpoints: 
\code{\link{get_chatters}()}
}
\concept{Hidden Endpoints}
