% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_request.R
\name{make_request}
\alias{make_request}
\title{Make a Twitch API Request}
\usage{
make_request(end_point, ..., clean_json = TRUE)
}
\arguments{
\item{end_point}{A character. The endpoint to make a request to.}

\item{...}{Named elements that will serve as parameters that will be used during the \code{httr::GET()} request.}

\item{clean_json}{A logical. If \code{TRUE}, clean and tidy the data. If \code{FALSE}, return the result of \code{httr::content}.}
}
\description{
Make a Twitch API Request
}
