% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%shh\%}
\alias{\%shh\%}
\title{Silence the Haters}
\usage{
lhs \%shh\% rhs
}
\arguments{
\item{lhs}{A value or the magrittr placeholder.}

\item{rhs}{A function call using the magrittr semantics.}
}
\description{
For situations when you want to mute \strong{known} warnings or messages in a {dplyr} chain.
}
