% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_categories}
\alias{search_categories}
\title{Search for Games or Categories}
\usage{
search_categories(query = NULL, first = NULL, after = NULL, clean_json = TRUE)
}
\arguments{
\item{query}{A character.     URl encoded search query.}

\item{first}{A numeric. Maximum number of objects to return. Maximum: 100. Default: 20.}

\item{after}{Cursor for forward pagination: tells the server where to start fetching the next set of results, in a multi-page response. The cursor value specified here is from the pagination response field of a prior query.}

\item{clean_json}{A logical. If \code{TRUE}, clean and tidy the data. If \code{FALSE}, return the result of \code{httr::content}.}
}
\description{
Returns a list of games or categories that match the query via name either entirely or partially.
}
\references{
https://dev.twitch.tv/docs/api/reference#search-categories
}
\seealso{
Other Search: 
\code{\link{search_channels}()}
}
\concept{Search}
