% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{tag_cleaner}
\alias{tag_cleaner}
\title{Extract and Clean Search Channel Tags}
\usage{
tag_cleaner(data_raw_element)
}
\arguments{
\item{data_raw_element}{An element from data_raw/response_content}
}
\value{
A list of tags per searched channel
}
\description{
Extract and Clean Search Channel Tags
}
