% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitch_auth.R
\name{twitch_auth}
\alias{twitch_auth}
\title{Get Auth Token}
\usage{
twitch_auth(
  client_id = Sys.getenv("TWITCH_CLIENT_ID"),
  client_secret = Sys.getenv("TWITCH_SECRET")
)
}
\arguments{
\item{client_id}{A character. Twitch App ID.}

\item{client_secret}{A character. Twitch App Secret.}
}
\value{
Invisible authorization data.
}
\description{
Get started with the Twitch API \href{https://dev.twitch.tv/docs/api/}{here}. \code{TWITCH_CLIENT_ID} and \code{TWITCH_SECRET} need to be set in the \emph{.Renviron} file. Set these values in \emph{.Renviron} with \code{usethis::edit_r_environ()}; ex: \code{TWITCH_CLIENT_ID=abc123}.
}
\examples{
\dontrun{
library(twitchr)

twitch_auth()
}
}
\concept{Authenticate}
