% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_stream_tags}
\alias{clean_stream_tags}
\title{Clean Stream Tags}
\usage{
clean_stream_tags(response_content, filter_language)
}
\arguments{
\item{response_content}{Content from a HTTP Request. This is expecting to take the output from \code{httr::content()}.}

\item{filter_language}{A character. Language to filter the return of tags.}
}
\value{
Clean and tidy tibble.
}
\description{
Clean Stream Tags
}
