% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clips.R
\name{get_all_clips}
\alias{get_all_clips}
\title{Get All Clips}
\usage{
get_all_clips(
  broadcaster_id = NULL,
  game_id = NULL,
  ended_at = NULL,
  started_at = NULL
)
}
\arguments{
\item{broadcaster_id}{A character. ID of the broadcaster for whom clips are returned. The number of clips returned is determined by the first query-string parameter (default: 20). Results are ordered by view count.}

\item{game_id}{A character. ID of the game for which clips are returned. The number of clips returned is determined by the first query-string parameter (default: 20). Results are ordered by view count.}

\item{ended_at}{A character. Ending date/time for returned clips, in RFC3339 format. (Note that the seconds value is ignored.) If this is specified, started_at also must be specified; otherwise, the time period is ignored.}

\item{started_at}{A character. Starting date/time for returned clips, in RFC3339 format. (The seconds value is ignored.) If this is specified, ended_at also should be specified; otherwise, the ended_at date/time will be 1 week after the started_at value.}
}
\description{
Gets clip information by broadcaster ID (one only), or game ID (one only). This function will handle pagination with \code{get_clips} automatically.
}
\references{
https://dev.twitch.tv/docs/api/reference#get-clips
}
\seealso{
Other Clips: 
\code{\link{get_clips}()}
}
\concept{Clips}
