% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{get_all_stream_tags}
\alias{get_all_stream_tags}
\title{Get All Stream Tags}
\usage{
get_all_stream_tags(
  after = NULL,
  first = NULL,
  tag_id = NULL,
  clean_json = TRUE
)
}
\arguments{
\item{after}{A character. Cursor for forward pagination: tells the server where to start fetching the next set of results, in a multi-page response. The cursor value specified here is from the pagination response field of a prior query.}

\item{first}{A numeric. Maximum number of objects to return. Maximum: 100. Default: 20.}

\item{tag_id}{A character. ID of a tag. Multiple IDs can be specified, separated by ampersands. If provided, only the specified tag(s) is(are) returned. Maximum of 100.}

\item{clean_json}{A logical. If \code{TRUE}, clean and tidy the data. If \code{FALSE}, return the result of \code{httr::content}.}
}
\description{
Get All Stream Tags
}
\references{
https://dev.twitch.tv/docs/api/reference#get-all-stream-tags
}
\seealso{
Other Tags: 
\code{\link{get_stream_tags}()}
}
\concept{Tags}
