% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidden_endpoints.R
\name{get_chatters}
\alias{get_chatters}
\title{Get a broadcasters current chatters}
\usage{
get_chatters(broadcaster)
}
\arguments{
\item{broadcaster}{A character. The name of the broadcaster.}
}
\value{
A tibble data frame of chatter data.
}
\description{
\strong{This uses undocumented API endpoints. This may change or break at anytime. Use at your own risk!}.
}
\examples{
\dontrun{
library(twitchr)

twitch_auth()

chatters <- get_chatters("KowAndToilet")
}
}
\references{
\url{https://thomassen.sh/twitch-api-endpoints/#tmi.twitch.tv/group/user/{USERNAME}/chatters}
}
\concept{Hidden Endpoints}
