% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clips.R
\name{get_clips}
\alias{get_clips}
\title{Get Clips}
\usage{
get_clips(
  broadcaster_id = NULL,
  game_id = NULL,
  id = NULL,
  after = NULL,
  before = NULL,
  ended_at = NULL,
  first = NULL,
  started_at = NULL,
  clean_json = TRUE
)
}
\arguments{
\item{broadcaster_id}{A character. ID of the broadcaster for whom clips are returned. The number of clips returned is determined by the first query-string parameter (default: 20). Results are ordered by view count.}

\item{game_id}{A character. ID of the game for which clips are returned. The number of clips returned is determined by the first query-string parameter (default: 20). Results are ordered by view count.}

\item{id}{A character. ID of the clip being queried. Limit: 100.}

\item{after}{A character. Cursor for forward pagination: tells the server where to start fetching the next set of results, in a multi-page response. This applies only to queries specifying broadcaster_id or game_id. The cursor value specified here is from the pagination response field of a prior query.}

\item{before}{A character. Cursor for backward pagination: tells the server where to start fetching the next set of results, in a multi-page response. This applies only to queries specifying broadcaster_id or game_id. The cursor value specified here is from the pagination response field of a prior query.}

\item{ended_at}{A character. Ending date/time for returned clips, in RFC3339 format. (Note that the seconds value is ignored.) If this is specified, started_at also must be specified; otherwise, the time period is ignored.}

\item{first}{A numeric. Maximum number of objects to return. Maximum: 100. Default: 20.}

\item{started_at}{A character. Starting date/time for returned clips, in RFC3339 format. (The seconds value is ignored.) If this is specified, ended_at also should be specified; otherwise, the ended_at date/time will be 1 week after the started_at value.}

\item{clean_json}{A logical. If \code{TRUE}, clean and tidy the data. If \code{FALSE}, return the result of \code{httr::content}.}
}
\value{
A tibble data frame of clip data.
}
\description{
Gets clip information by clip ID (one or more), broadcaster ID (one only), or game ID (one only). Note: The clips service returns a maximum of 1000 clips. For a query to be valid, \code{id} (one or more), \code{broadcaster_id}, or \code{game_id} must be specified. You may specify only one of these parameters.
}
\examples{
\dontrun{
twitch_auth()

user <- get_users(login = "KowAndToilet")

clips <- get_clips(broadcaster_id = user$id)
}
}
\references{
https://dev.twitch.tv/docs/api/reference#get-clips
}
\seealso{
Other Clips: 
\code{\link{get_all_clips}()}
}
\concept{Clips}
