% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{get_stream_tags}
\alias{get_stream_tags}
\title{Get Stream Tags}
\usage{
get_stream_tags(broadcaster_id = NULL, language = "en-us", clean_json = TRUE)
}
\arguments{
\item{broadcaster_id}{A character. ID of the stream thats tags are going to be fetched.}

\item{language}{A character. Wanted language. Need to list languages here...}

\item{clean_json}{A logical. If \code{TRUE}, clean and tidy the data. If \code{FALSE}, return the result of \code{httr::content}.}
}
\description{
Gets the list of tags for a specified stream (channel).
}
\references{
https://dev.twitch.tv/docs/api/reference#get-stream-tags
}
\seealso{
Other Tags: 
\code{\link{get_all_stream_tags}()}
}
\concept{Tags}
