% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitchr-package.R
\docType{package}
\name{twitchr-package}
\alias{twitchr}
\alias{twitchr-package}
\title{twitchr: R Interface for the Twitch API}
\description{
R Wrapper for the Twitch API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/KoderKow/twitchr}
  \item Report bugs at \url{https://github.com/KoderKow/twitchr/issues}
}

}
\author{
\strong{Maintainer}: Kyle Harris \email{koderkow@gmail.com}

}
\keyword{internal}
